[backport] Carve up Types.scala
Step one of a plan to bring some order and thread
safety to this neck of the woods.

More info:

  https://gist.github.com/retronym/5081754
  https://groups.google.com/forum/?fromgroups=#!topic/scala-internals/MOvmcnbyb_g

Note that the sub package is named 'tpe' and not 'Types' to avoid
problems on case insensitive file systems, given the existing
trait 'Types.'

============================================================

Some contortions were required to avoid binary compatibility
breakages with `scala.reflect.runtime` (ironically via the
SynchronizedXxx wrappers, which don't solve reflection safety,
and now slightly impede the real solution.)

MiMa fails when I run it locally in a different part of the
standard library:

    bc.run:
         [java] Found 1 binary incompatibiities (2 were filtered out)
         [java] =====================================================
         [java]  * synthetic method
         [java]    scala$xml$dtd$ElementValidator$$find$2(java.lang.String,scala.Option)Boolean
         [java]    in class scala.xml.dtd.ElementValidator's type has changed; was
         [java]    (java.lang.String,scala.Option)Boolean, is now:
         [java]    (java.lang.String,scala.collection.Seq)Boolean
         [java] Generated filter config definition

I'm not sure if whether to trust that, so we'll see how this commit
goes on Jenkins.
